synq: fix build-web-playground Perfetto dialect WASM compilation
Three bugs prevented syntaqlite-perfetto.wasm from being built:

1. _common.y hardcoded "csrc/sqlite/dialect_builder.h" which is the
   SQLite-internal layout path. For external/amalgamation builds the
   file lives at "csrc/dialect_builder.h", so the amalgamator could
   not resolve it and stripped it — leaving all synq_parse_* builder
   functions undeclared. Fix: replace with @DIALECT_BUILDER_H@
   placeholder and substitute the correct layout path in codegen_api.rs
   before lemon runs.

2. sqlite_parse.h declared SynqXxxParse with uint32_t token_type but
   lemon generates int yymajor. In the amalgamation both appear in the
   same TU, causing "conflicting types". Fix: declare the function with
   int, update the grammar.h parser_feed function pointer to int, and
   cast uint32_t→int at the SYNQ_PARSER_FEED call site in parser.c
   (matching the existing pattern in token_wrapped.c).

3. syntaqlite-cli/src/lib.rs lacked #![allow(missing_docs)] after the
   Cargo.toml lints override was removed, producing a spurious warning.

+++ syntaqlite-syntax/csrc/parser.c
@@ -125,7 +125,7 @@ static int feed_one_token(SyntaqliteParser* p,
                           uint32_t token_idx) {
   SynqParseToken minor = {
       .z = text, .n = len, .type = token_type, .token_idx = token_idx};
-  SYNQ_PARSER_FEED(p->grammar.tmpl, p->lemon, token_type, minor, &p->ctx);
+  SYNQ_PARSER_FEED(p->grammar.tmpl, p->lemon, (int)token_type, minor, &p->ctx);
   p->last_token_type = token_type;
 
   if (p->ctx.error) {
